Fix nil regions
When creating a Physical::Location with a country code and a region
code, the region would be `nil` even if the region code was a region
in the specified country. This commit fixes that behaviour by explicitly
checking whether the `country` instance variable is a valid `Carmen::Country`.
Previously, we checked for the `country` name, which in the initialization
context does not hit the getter, but the named argument to the `initialize`
method.