Do not sort dimensions for measuring
When a user sends the dimensions, we take the first dimension as length,
the second as width, and the third as height. We document this, and thus
now have support for packages that are taller than wide, for example,
but cannot be turned around.

This commit also removes all 'convenience' methods that further increase
the confusion about what exactly constitutes the 'lenght', 'width' or
'height' of a package. We especially eschew the notion of 'depth',
because that word can either mean 'how deep' (how high from the ground)
or 'how far away from the front side'.