Delegate package#volume to its container
In the scenario were a package contains other packages the
remaining_volume calculation breaks, because it depends on the
items (in this case another `Physical::Package`) volume.

Delegating the package volume to its container volume to fix this.

Also it is convinient to have the packages (outer-)volume as well.