Alias inner dimensions for `Pallet` class
This adds aliases for inner dimensions to the `Pallet` class. The
`Package` class calls `#inner_volume` when calculating remaining
volume, and since a `Pallet` is intended to be used as the container
for a `Package`, we need to respond to the method even though a
`Pallet` doesn't actually have inner dimensions.