Physical::Package Items Value
Some services that take a physical package want to know the value of the contents of the package. Aggregate the cost of the items and return for #items_value. Cost however is optional. When it doesn't exist Item#cost returns nil. If any item doesn't have a cost return nil for Package#items_value so we don't provide partial value