Update `Physical::Shipment` to accept structures
Instead of passing pallets to `Physical::Shipment` we should be
passing structures instead. Structures could be pallets, skids, or
anything else that contains one or more packages.

The idea is to support modeling shipments containing one or more
pallets (each of which holds one or more packages) in addition to
separate packages that aren't being shipped on a pallet.

This commit deprecates the `pallets` argument.